add inheritance for 'task.js' && 'task-edit.js'
